Added function generate_boundary_input and som auxiliry functions. It produced the same result as the original code when reading NORA3 boundary spectra for 2019-01-09T00:00:00 to 2019-01-10T00:00:00 and writing the SWAN asciis (checked that they are same with diff).
This new functions should save some repetitive code, since it should be directly applicable to the 4 km operational WAM (not tested) and also the WW3 output to netcdf is implementable as a separate function using the same list of xarray datasets.

NB! The new function doesn't take all inputs, and some are hardcoded (for now) in the function.