Changes to the output file workflow:
1) Added possibility to set file names using regular expressions, e.g. 
"$Grid_$T0-$T1'

2) Added defaul regular expressions and datestrings to a module 
defaults.py. These are used in case nothing is specified by the user, 
and they are defined separately for each object and model.

3) The Forcing and Boundary objects now remember to which file and which 
path they have been written.

4) The information about the output files is used by the fucntion 
creating the input file when the objects are provided. If objects are 
provided that has not been written to a file, then the default file 
names of default.py are used. However, all automatical name generation 
can be overridden by explicitly providing a filestring to the function. 
The filestring can contain regular expressions, but can also be just a 
plain hard coded file name.

5) The function writing the input file for SWAN also reads the start and 
end time of the model run directly from the Forcing and Boundary 
objects. Again, this can be overridden by providing an explicit start 
and end time.