Bug fix in spectral processor
There was a bug in the flip_spec function that made it work incorrectly for some specific transformations (specific orientiation of the directional vector and directional resolution). This bug has been fixed and unit tests have been expanded.

The boundary reader for reading WW3 netcdf-files was corrected for a bug in reading the directions from the xarray dataset and some function handles were harmonized with the rest of the code.